refactor(auth): moved email password apis into provider
